Intermediate check-in. Not running (asserts on NState remove via GUI).
• NStateMachine/Map/State is now serializing properly (needed public members, all elements serializable/notEvilStructs, and not doing any kind of init other than the declaration line itself).
• NStateMachineEditor now has primative state-showing (broken currently).

Problem: It seems to be finding a state to remove, then going to delete that state and it doens't exist. I suspect I'll need to either go fully "string-duck" ("if it says it's name is the name we're looking for, then that's good enough") or fully "OO" ("they must give us the exact object they want to remove, and it will only be remove if we find the exact object"). I suspect the latter would be ideal if I were using ScriptableObjects, but the former may work better for objects that know how to garbage collect nicely.